Fix error when client did not send Content-Lenght
In this situation we assume content lenght is 0.

+++ src/HTTPPacket.hpp
@@ -12,12 +12,12 @@ public:
 	void save_line(std::string line);
 	void save_json(std::string json);
 	std::string get_json();
-	size_t get_content_length();
+	unsigned int get_content_length();
 	bool is_valid_request();
 private:
 	std::string json;
 	std::vector<std::pair<std::string, std::string>> headers;
-	int contentLength;
+	unsigned int contentLength;
 	bool valid_request;
 };
